For #77 .

@Stebalien @raulk 

- Pre-calculate handshake message lengths so we can go to the pool for them.
-  Refactor the benchmarking code to test with read and write buffers of different lengths.
- Respect Context Cancellation.
- First read length of message -> see if we have a large enough buffer -> zero alloc read and decrypt.
- Don't allocate twice when sending/writing messages.

Even after all these optimisations, while there is an improvement in the `ns/op` (I suspect because we don't go the pool often) and the bytes used, the numbers of allocs isn't changing meaningfully. Even for 0 alloc cases where the buffer we pass for read should be used for reading the message and then decrypting in place, the number of allocs is equal or roughly the same. I'm trying to debug why this is the case and if I'm doing something wrong somewhere.

_Yet to come up with an abstraction for sending/writing messages. Will create an issue to track it._

```
aarshshah@Aarshs-MacBook-Pro go-libp2p-noise % benchcmp oldfinal.txt newfinal.txt 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                      old ns/op      new ns/op      delta
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-12         1894036        1886413        -0.40%
BenchmarkTransfer1MB/readBuffer_>_decrypted_plaintext-12       1890222        1869960        -1.07%
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-12       1905608        1879583        -1.37%
BenchmarkTransfer100MB/readBuffer_>_decrypted_plaintext-12     217205716      213775674      -1.58%
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-12     220203489      216048322      -1.89%
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-12       216406756      211674437      -2.19%
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-12     1476738831     1462894701     -0.94%
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-12       1464702875     1430231886     -2.35%
BenchmarkTransfer500Mb/readBuffer_>_decrypted_plaintext-12     1463074671     1447557381     -1.06%
BenchmarkHandshakeXX-12                                        704583         702165         -0.34%

benchmark                                                      old allocs     new allocs     delta
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-12         144            143            -0.69%
BenchmarkTransfer1MB/readBuffer_>_decrypted_plaintext-12       144            144            +0.00%
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-12       144            144            +0.00%
BenchmarkTransfer100MB/readBuffer_>_decrypted_plaintext-12     23409          23407          -0.01%
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-12     23409          23407          -0.01%
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-12       23408          22797          -2.61%
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-12     328601         328594         -0.00%
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-12       328595         328568         -0.01%
BenchmarkTransfer500Mb/readBuffer_>_decrypted_plaintext-12     328601         328594         -0.00%
BenchmarkHandshakeXX-12                                        298            293            -1.68%

benchmark                                                      old bytes     new bytes     delta
BenchmarkTransfer1MB/readBuffer_>_encrypted_message-12         4432          4422          -0.23%
BenchmarkTransfer1MB/readBuffer_>_decrypted_plaintext-12       4200          4068          -3.14%
BenchmarkTransfer1MB/readBuffer_<_decrypted_plaintext-12       4352          4262          -2.07%
BenchmarkTransfer100MB/readBuffer_>_decrypted_plaintext-12     389746        388447        -0.33%
BenchmarkTransfer100MB/readBuffer_<_decrypted_plaintext-12     393645        388026        -1.43%
BenchmarkTransfer100MB/readBuffer_>_encrypted_message-12       392669        372540        -5.13%
BenchmarkTransfer500Mb/readBuffer_<_decrypted_plaintext-12     5292584       5283650       -0.17%
BenchmarkTransfer500Mb/readBuffer_>_encrypted_message-12       5290774       5270556       -0.38%
BenchmarkTransfer500Mb/readBuffer_>_decrypted_plaintext-12     5292608       5290948       -0.03%
BenchmarkHandshakeXX-12                                        23258         22364         -3.84%
```